Voice-Command-Shopping-Assistant/requirements.txt
fastapi
uvicorn
assemblyai
python-dotenv
requests
pymongo
sentence-transformers
faiss-cpu
rapidfuzz
